requirements.txt
# 单词自动化助手依赖项
# 使用国内镜像源加速安装：
# pip install -r requirements.txt -i https://pypi.tuna.tsinghua.edu.cn/simple

kivy>=2.1.0
opencv-python>=4.5.0
requests>=2.25.0
pillow>=8.0.0
appium-python-client>=2.0.0
pytesseract>=0.3.8
speechrecognition>=3.8.1
numpy>=1.20.0
